--- src/main/java/com/sixshop/sixspace/vacation/repository/VacationRepository2.java
@@ -1,28 +1,0 @@
-package com.sixshop.sixspace.vacation.repository;
-
-import static com.sixshop.sixspace.vacation.domain.QVacation.vacation;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import com.sixshop.sixspace.vacation.domain.Vacation;
-import java.time.LocalDateTime;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
-@RequiredArgsConstructor
-@Repository
-public class VacationRepository2 {
-
-    private final JPAQueryFactory queryFactory;
-
-    public List<Vacation> findAllByBetweenDates(LocalDateTime start, final LocalDateTime end) {
-        return queryFactory
-            .select(vacation)
-            .from(vacation)
-            .where(vacation.startDateTime.between(start, end)
-                .or(vacation.endDateTime.between(start, end)))
-            .orderBy(vacation.startDateTime.asc())
-            .fetch();
-    }
-
-}
